Renaming traversers to analyzers
