+++ packages/core/src/analyzers/json-traverser.ts
